fix(async): move AsyncQueue to separate class, add ServiceContainer::has(), fix test
